Modify folder structure to Domain-Driven Disign
