+++ lib/core/theme/app_theme.dart
@@ -1,0 +1,39 @@
+import 'package:flutter/material.dart';
+
+class AppTheme {
+  static final ThemeData lightTheme = ThemeData(
+    brightness: Brightness.light,
+    primaryColor: Colors.blue,
+    hintColor: Colors.blueAccent,
+    scaffoldBackgroundColor: Colors.white,
+    textTheme: const TextTheme(
+      bodyLarge: TextStyle(color: Colors.black),
+      bodyMedium: TextStyle(color: Colors.black54),
+    ),
+    appBarTheme: const AppBarTheme(
+      color: Colors.blue,
+      iconTheme: IconThemeData(
+        color: Colors.white,
+        size: 20,
+      ),
+    ),
+  );
+
+  static final ThemeData darkTheme = ThemeData(
+    brightness: Brightness.dark,
+    primaryColor: Colors.blueGrey,
+    hintColor: Colors.cyanAccent,
+    scaffoldBackgroundColor: Colors.black,
+    textTheme: const TextTheme(
+      bodyLarge: TextStyle(color: Colors.white),
+      bodyMedium: TextStyle(color: Colors.white70),
+    ),
+    appBarTheme: const AppBarTheme(
+      color: Colors.blueGrey,
+      iconTheme: IconThemeData(
+        color: Colors.white,
+        size: 20,
+      ),
+    ),
+  );
+}
