+++ lib/core/validators/auth_validator.dart
@@ -1,0 +1,45 @@
+import 'package:email_validator/email_validator.dart';
+
+class AuthValidator {
+  static String? validateUsername(String? value) {
+    if (value == null || value.isEmpty) {
+      return 'Username is required';
+    }
+
+    return null;
+  }
+
+  static String? validateEmail(String? value) {
+    if (value == null || value.isEmpty) {
+      return 'Email is required';
+    }
+
+    if (!EmailValidator.validate(value)) {
+      return 'Invalid email';
+    }
+
+    return null;
+  }
+
+  static String? validatePassword(String? value) {
+    if (value == null || value.isEmpty) {
+      return 'Password is required';
+    }
+
+    if (value.length < 6) {
+      return 'Password must be at least 6 characters';
+    }
+
+    return null;
+  }
+
+  static String? validateConfirmPassword(
+      String? confirmPassword, String password) {
+    if (confirmPassword == null || confirmPassword.isEmpty) {
+      return 'Please enter your password confirmation';
+    } else if (confirmPassword != password) {
+      return "Password doesn't match.";
+    }
+    return null;
+  }
+}
