+++ lib/data/data.dart
@@ -1,0 +1,54 @@
+
+import 'package:myapp/data/models/coffee_model.dart';
+
+final Map<String, CoffeeItem> coffeeDictionary = {
+    '1': CoffeeItem(
+      name: 'Colombia Supremo',
+      category: 'Cappucino',
+      image: 'assets/images/coffee2.jpg',
+      description: 'A smooth and chocolatey coffee from Colombia',
+      price: 15.99,
+    ),
+    '2': CoffeeItem(
+      name: 'Ethiopian Yirgacheffe',
+      category: 'Machiato',
+      image: 'assets/images/coffee3.jpg',
+      description: 'A floral and citrusy coffee from Ethiopia',
+      price: 18.99,
+    ),
+    '3': CoffeeItem(
+      name: 'Guatemalan Huehuetenango',
+      category: 'Machiato',
+      image: 'assets/images/coffee4.jpg',
+      description: 'A spicy and caramel-like coffee from Guatemala',
+      price: 16.99,
+    ),
+    '4': CoffeeItem(
+      name: 'Kenyan AA',
+      category: 'Latte',
+      image: 'assets/images/coffee5.webp',
+      description: 'A vibrant and fruity coffee from Kenya',
+      price: 19.99,
+    ),
+    '5': CoffeeItem(
+      name: 'Cappucino',
+      category: 'Dark Roast',
+      image: 'assets/images/coffee6.jpg',
+      description: 'An earthy and syrupy coffee from Indonesia',
+      price: 17.99,
+    ),
+    '6': CoffeeItem(
+      name: 'Kenyan AA',
+      category: 'latte',
+      image: 'assets/images/coffee7.jpg',
+      description: 'A vibrant and fruity coffee from Kenya',
+      price: 19.99,
+    ),
+    '7': CoffeeItem(
+      name: 'Ethiopian Yirgacheffe',
+      category: 'Machiato',
+      image: 'assets/images/coffee3.jpg',
+      description: 'A floral and citrusy coffee from Ethiopia',
+      price: 18.99,
+    ),
+  };
