+++ lib/main.dart
@@ -1,19 +1,13 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:go_router/go_router.dart';
 import 'package:myapp/firebase_options.dart';
-import 'package:myapp/src/presentation/auth/bloc/auth_bloc.dart';
-import 'package:myapp/src/presentation/auth/repositories/auth_repository.dart';
-import 'package:myapp/src/presentation/cart/bloc/cart_bloc.dart';
-import 'package:myapp/src/presentation/auth/login.dart';
-import 'package:myapp/src/presentation/auth/register.dart';
-import 'package:myapp/src/presentation/cart/cart.dart';
-import 'package:myapp/src/presentation/favorites/bloc/favorites_bloc.dart';
-import 'package:myapp/src/presentation/favorites/favorites.dart';
-import 'package:myapp/src/presentation/home/home.dart';
-import 'package:myapp/src/presentation/orders/orders.dart';
-import 'package:myapp/src/presentation/welcome/welcome.dart';
+import 'package:myapp/presentation/blocs/auth/auth_bloc.dart';
+import 'package:myapp/domain/repositories/auth_repository.dart';
+import 'package:myapp/presentation/blocs/cart/cart_bloc.dart';
+import 'package:myapp/presentation/blocs/favorites/favorites_bloc.dart';
+import 'package:myapp/core/routes/app_route.dart';
+
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -43,7 +37,7 @@ class MyApp extends StatelessWidget {
         ),
       ],
       child: MaterialApp.router(
-        routerConfig: _router,
+        routerConfig: router,
         title: 'Our wonderful Ethiopian Coffee shop',
         theme: ThemeData(
           useMaterial3: true,
@@ -53,58 +47,3 @@ class MyApp extends StatelessWidget {
     );
   }
 }
-
-final GoRouter _router = GoRouter(
-  initialLocation: '/welcome',
-  routes: <RouteBase>[
-    GoRoute(
-      name: '/welcome',
-      path: '/welcome',
-      builder: (context, state) {
-        return const WelcomePage();
-      },
-    ),
-    GoRoute(
-      name: '/login',
-      path: '/login',
-      builder: (context, state) {
-        return const LoginPage();
-      },
-    ),
-    GoRoute(
-      name: '/register',
-      path: '/register',
-      builder: (context, state) {
-        return const RegisterPage();
-      },
-    ),
-    GoRoute(
-      name: '/home',
-      path: '/home',
-      builder: (context, state) {
-        return const HomePage();
-      },
-    ),
-    GoRoute(
-      name: '/favorites',
-      path: '/favorites',
-      builder: (context, state) {
-        return const FavoritesList();
-      },
-    ),
-    GoRoute(
-      name: '/orders',
-      path: '/orders',
-      builder: (context, state) {
-        return const OrdersList();
-      },
-    ),
-    GoRoute(
-      name: '/cart',
-      path: '/cart',
-      builder: (context, state) {
-        return const AddToCart();
-      },
-    ),
-  ],
-);
