+++ lib/presentation/blocs/auth/auth_bloc.dart
@@ -1,0 +1,49 @@
+import 'package:bloc/bloc.dart';
+import 'package:equatable/equatable.dart';
+import 'package:myapp/domain/repositories/auth_repository.dart';
+import 'package:myapp/data/models/user_model.dart';
+
+part 'auth_event.dart';
+part 'auth_state.dart';
+
+class AuthBloc extends Bloc<AuthEvent, AuthState> {
+  final AuthRepository authRepository;
+
+  AuthBloc({required this.authRepository}) : super(AuthInitial()) {
+    on<SignUpRequested>((event, emit) async {
+      emit(AuthLoading());
+      try {
+        await authRepository.signUp(email: event.email, password: event.password);
+        final user = authRepository.currentUser;
+        if (user != null) {
+          emit(AuthAuthenticated(user: user));
+        } else {
+          emit(AuthUnauthenticated());
+        }
+      } catch (e) {
+        emit(AuthError(e.toString()));
+      }
+    });
+
+    on<LoginRequested>((event, emit) async {
+      emit(AuthLoading());
+      try {
+        await authRepository.login(email: event.email, password: event.password);
+        final user = authRepository.currentUser;
+        if (user != null) {
+          emit(AuthAuthenticated(user: user));
+        } else {
+          emit(AuthUnauthenticated());
+        }
+      } catch (e) {
+        emit(AuthError(e.toString()));
+      }
+    });
+
+    on<SignOutRequested>((event, emit) async {
+      emit(AuthLoading());
+      await authRepository.signOut();
+      emit(AuthUnauthenticated());
+    });
+  }
+}
