+++ lib/presentation/blocs/auth/auth_event.dart
