+++ lib/presentation/blocs/auth/auth_state.dart
