+++ lib/presentation/blocs/favorites/favorites_event.dart
