+++ lib/presentation/blocs/favorites/favorites_state.dart
