+++ lib/presentation/widgets/inputs/custom_text_input.dart
@@ -1,0 +1,49 @@
+import 'package:flutter/material.dart';
+
+class CustomTextInput extends StatefulWidget {
+  final String labelText;
+  final String hintText;
+  final TextEditingController controller;
+  final TextInputType keyboardType;
+  final bool obscureText;
+  final FormFieldValidator<String>? validator;
+  final ValueChanged<String>? onChanged;
+  final Widget? suffixIcon;
+
+  const CustomTextInput({
+    super.key,
+    required this.labelText,
+    required this.hintText,
+    required this.controller,
+    this.keyboardType = TextInputType.text,
+    this.obscureText = false,
+    this.validator,
+    this.onChanged,
+    this.suffixIcon,
+  });
+
+  @override
+  State<CustomTextInput> createState() => _CustomTextInputState();
+}
+
+class _CustomTextInputState extends State<CustomTextInput> {
+  @override
+  Widget build(BuildContext context) {
+    return ConstrainedBox(
+      constraints: const BoxConstraints(maxWidth: 500),
+      child: TextFormField(
+        controller: widget.controller,
+        keyboardType: widget.keyboardType,
+        obscureText: widget.obscureText,
+        validator: widget.validator,
+        decoration: InputDecoration(
+          labelText: widget.labelText,
+          hintText: widget.hintText,
+          border: const OutlineInputBorder(),
+          suffixIcon: widget.suffixIcon,
+        ),
+        onChanged: widget.onChanged,
+      ),
+    );
+  }
+}
