--- lib/src/presentation/cart/cart.dart
@@ -1,100 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:myapp/src/presentation/cart/bloc/cart_bloc.dart';
-import 'package:myapp/src/common/footer.dart';
-import 'package:myapp/src/presentation/detail/detail.dart';
-
-class AddToCart extends StatefulWidget {
-  const AddToCart({super.key});
-
-  @override
-  AddToCartState createState() => AddToCartState();
-}
-
-class AddToCartState extends State<AddToCart> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Colors.brown,
-        title: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          children: [
-            const Text(
-              'My Cart',
-              style: TextStyle(fontSize: 18, color: Colors.white),
-            ),
-            IconButton(
-                icon: const Icon(
-                  Icons.shopping_cart,
-                  color: Colors.white,
-                ),
-                onPressed: () {})
-          ],
-        ),
-      ),
-      body: BlocBuilder<CartBloc, CartState>(
-        builder: (context, state) {
-          if (state is CartUpdated) {
-            final cartItems = state.cartItems;
-            if (cartItems.isEmpty) {
-              return const Center(
-                child: Text('No items in cart!'),
-              );
-            } else {
-              return ListView.builder(
-                itemCount: cartItems.length,
-                itemBuilder: (context, index) {
-                  final item = cartItems[index];
-                  return Card(
-                    child: GestureDetector(
-                      onTap: () {
-                        Navigator.push(
-                          context,
-                          MaterialPageRoute(
-                              builder: (context) =>
-                                  DetailPage(coffeeItem: item)),
-                        );
-                      },
-                      child: ListTile(
-                        leading: ClipRRect(
-                          borderRadius: BorderRadius.circular(10),
-                          child: Image.asset(
-                            item.image,
-                            width: 50,
-                            height: 50,
-                            fit: BoxFit.cover,
-                          ),
-                        ),
-                        title: Text(item.name,
-                            style: const TextStyle(color: Colors.brown)),
-                        subtitle: Text('\$ ${item.price.toString()}',
-                            style: const TextStyle(color: Colors.brown)),
-                        trailing: IconButton(
-                          icon: const Icon(Icons.delete,
-                              color: Colors.red, size: 18),
-                          onPressed: () {
-                            BlocProvider.of<CartBloc>(context).add(
-                              RemoveItemFromCart(item),
-                            );
-                          },
-                        ),
-                      ),
-                    ),
-                  );
-                },
-              );
-            }
-          } else {
-            return const Center(
-              child: Text('No items in cart!'),
-            );
-          }
-        },
-      ),
-      bottomNavigationBar: const BottomNavigation(
-        currentIndex: 3,
-      ),
-    );
-  }
-}
